requirements.txt with Django==3.2a1 and prerequisits after running pip install --pre django
This is an alpha release, I hope I don't regret it.
The good news is that this is the next LTS release, so it's going to be
supported for a long time.
This is also the first LTS release to require Python 3.6, a good thing
as there were some significant features in that release