fix issue #2 migration 0001_initial.py fails to apply with database set to MySQL due to excessive number of CHAR fields
-------
remove a bunch of fields from a model that will be deleted in another migration
these were purged because there were too many CHAR fields for a mysql row
mysql requires varchar etc when using so many fields in a row
-------

issue text:
migration jeromeneedsill/illrequest/migrations/0001_initial.py fails to apply with database set to MySQL due to excessive number of CHAR fields.

These are on the database table illrequest which we end up getting rid of in a later migration, jeromeneedsill/illrequest/migrations/0004_delete_illrequest.py

As such, the solution is to just delete the excessive fields from jeromeneedsill/illrequest/migrations/0001_initial.py as if they never existed in the first place. The net effect for new deployments is the table ends up added with less fields and then deleted by a later migration. This is better than deleting the entire 0001_initial.py migration because there are already development/production systems with the history of that migration.

A patch has already been written and will be soon applied, closing this issue shortly after open. (a worthy practice because this was a bug in previous releases)